feat(caps): add agent and prompt capability modules
